requirements.txt
streamlit==1.41.0
pillow==11.0.0
openai==1.57.2
requests==2.32.3 